FAD-6340 moved save as csv button
