+++ src/components/collection/Collection.js
@@ -1,14 +1,12 @@
 /* eslint max-lines: ["error", 200] */
 import React, { Component } from 'react';
 import CollectionPropTypes from './Collection.propTypes';
-import { Button } from '@sparkpost/matchbox';
 import qs from 'query-string';
 import _ from 'lodash';
 import { withRouter } from 'react-router-dom';
 import Pagination, { defaultPerPageButtons } from './Pagination';
 import FilterBox from './FilterBox';
 import { objectSortMatch } from 'src/helpers/sortMatch';
-import Papa from 'papaparse';
 
 const PassThroughWrapper = (props) => props.children;
 const NullComponent = () => null;
@@ -105,7 +103,7 @@ export class Collection extends Component {
   }
 
   renderPagination() {
-    const { rows, perPageButtons, pagination } = this.props;
+    const { rows, perPageButtons, pagination, saveCsv } = this.props;
     const { currentPage, perPage, filteredRows } = this.state;
 
     if (!pagination || !currentPage) { return null; }
@@ -118,25 +116,11 @@ export class Collection extends Component {
         perPageButtons={perPageButtons}
         onPageChange={this.handlePageChange}
         onPerPageChange={this.handlePerPageChange}
+        saveCsv={saveCsv}
       />
     );
   }
 
-  formatToCsv = () => {
-    const { rows } = this.props;
-    const data = Papa.unparse(rows);
-    return `data:text/csv;charset=utf-8,${encodeURI(data)}`;
-  }
-
-  renderCSVSave() {
-    const { saveCsv } = this.props;
-
-    if (!saveCsv) { return null; }
-
-    const now = Math.floor(Date.now() / 1000);
-    return <Button download={`sparkpost-csv-${now}.csv`} to={this.formatToCsv()}>Save As CSV</Button>;
-
-  }
 
   render() {
     const {
@@ -162,7 +146,6 @@ export class Collection extends Component {
           </BodyWrapper>
         </OuterWrapper>
         {this.renderPagination()}
-        {this.renderCSVSave()}
       </div>
     );
   }
